refactor: move splitChildren to its own file
